refactor to add post controller
